Tag v1.7.0

+++ test/lib/OAuth2/Storage/Bootstrap.php
@@ -171,20 +171,29 @@ class Bootstrap
     public function getCouchbase()
     {
         if (!$this->couchbase) {
-            $skipCouchbase = $this->getEnvVar('SKIP_COUCHBASE_TESTS');
-            if (!$skipCouchbase && class_exists('Couchbase')) {
-
-                $couchbase = new \Couchbase(array('localhost:8091'), '', '', 'auth', false);
-                if ($this->testCouchbaseConnection($couchbase)) {
-                    $this->clearCouchbase($couchbase);
-                    $this->createCouchbaseDB($couchbase);
-
-                    $this->couchbase = new CouchbaseDB($couchbase);
-                } else {
-                    $this->couchbase = new NullStorage('Couchbase', 'Unable to connect to Couchbase server on "localhost:8091"');
-                }
-            } else {
+            if ($this->getEnvVar('SKIP_COUCHBASE_TESTS')) {
+                $this->couchbase = new NullStorage('Couchbase', 'Skipping Couchbase tests');
+            } elseif (!class_exists('Couchbase')) {
                 $this->couchbase = new NullStorage('Couchbase', 'Missing Couchbase php extension. Please install couchbase.so');
+            } else {
+                // round-about way to make sure couchbase is working
+                // this is required because it throws a "floating point exception" otherwise
+                $code = "new \Couchbase(array('localhost:8091'), '', '', 'auth', false);";
+                $exec = sprintf('php -r "%s"', $code);
+                $ret = exec($exec, $test, $var);
+                if ($ret != 0) {
+                    $couchbase = new \Couchbase(array('localhost:8091'), '', '', 'auth', false);
+                    if ($this->testCouchbaseConnection($couchbase)) {
+                        $this->clearCouchbase($couchbase);
+                        $this->createCouchbaseDB($couchbase);
+
+                        $this->couchbase = new CouchbaseDB($couchbase);
+                    } else {
+                        $this->couchbase = new NullStorage('Couchbase', 'Unable to connect to Couchbase server on "localhost:8091"');
+                    }
+                } else {
+                    $this->couchbase = new NullStorage('Couchbase', 'Error while trying to connect to Couchbase');
+                }
             }
         }
 
@@ -258,12 +267,13 @@ class Bootstrap
         ));
 
         $sys->create_column_family('oauth2_test', 'auth');
+        $cassandra = new \phpcassa\Connection\ConnectionPool('oauth2_test', array('127.0.0.1:9160'));
+        $cf = new \phpcassa\ColumnFamily($cassandra, 'auth');
 
         // populate the data
         $storage->setClientDetails("oauth_test_client", "testpass", "http://example.com", 'implicit password');
         $storage->setAccessToken("testtoken", "Some Client", '', time() + 1000);
         $storage->setAuthorizationCode("testcode", "Some Client", '', '', time() + 1000);
-        $storage->setUser("testuser", "password");
 
         $storage->setScope('supportedscope1 supportedscope2 supportedscope3 supportedscope4');
         $storage->setScope('defaultscope1 defaultscope2', null, 'default');
@@ -281,6 +291,13 @@ class Bootstrap
         $storage->setScope('clientscope3', 'Test Default Scope Client ID 2', 'default');
 
         $storage->setClientKey('oauth_test_client', $this->getTestPublicKey(), 'test_subject');
+
+        $cf->insert("oauth_public_keys:ClientID_One", array('__data' => json_encode(array("public_key" => "client_1_public", "private_key" => "client_1_private", "encryption_algorithm" => "RS256"))));
+        $cf->insert("oauth_public_keys:ClientID_Two", array('__data' => json_encode(array("public_key" => "client_2_public", "private_key" => "client_2_private", "encryption_algorithm" => "RS256"))));
+        $cf->insert("oauth_public_keys:", array('__data' => json_encode(array("public_key" => $this->getTestPublicKey(), "private_key" =>  $this->getTestPrivateKey(), "encryption_algorithm" => "RS256"))));
+
+        $cf->insert("oauth_users:testuser", array('__data' =>json_encode(array("password" => "password", "email" => "testuser@test.com", "email_verified" => true))));
+
     }
 
     private function createSqliteDb(\PDO $pdo)
@@ -309,6 +326,10 @@ class Bootstrap
 
     private function createPostgresDb()
     {
+        if (!`psql postgres -tAc "SELECT 1 FROM pg_roles WHERE rolname='postgres'"`) {
+            `createuser -s -r postgres`;
+        }
+
         `createdb -O postgres oauth2_server_php`;
     }
 
@@ -319,7 +340,9 @@ class Bootstrap
 
     private function removePostgresDb()
     {
-        `dropdb oauth2_server_php`;
+        if (trim(`psql -l | grep oauth2_server_php | wc -l`)) {
+            `dropdb oauth2_server_php`;
+        }
     }
 
     public function runPdoSql(\PDO $pdo)
